Delete girl_names.html
Day 20 exercise, speech recognition. I added a girls names recognition - If girls name is said the program will add a female unicode sign to the name. If the name is Sabina (my fiancee) it will put a heart unicode. I hooked up a list from wikipedia, which i had to transfer from unordered lists to an array of names.